Move temp_rh to hardware
