+++ src/vario/hardware/temp_rh.h
