+++ src/vario/power.cpp
@@ -5,10 +5,10 @@
 #include "hardware/Leaf_SPI.h"
 #include "hardware/configuration.h"
 #include "hardware/io_pins.h"
+#include "hardware/temp_rh.h"
 #include "instruments/baro.h"
 #include "instruments/gps.h"
 #include "instruments/imu.h"
-#include "instruments/temp_rh.h"
 #include "logging/log.h"
 #include "power.h"
 #include "storage/sd_card.h"
