+++ src/vario/ui/display/pages/primary/page_navigate.cpp
@@ -1,10 +1,10 @@
 #include <Arduino.h>
 #include <U8g2lib.h>
 
+#include "hardware/temp_rh.h"
 #include "instruments/baro.h"
 #include "instruments/gps.h"
 #include "instruments/imu.h"
-#include "instruments/temp_rh.h"
 #include "logging/log.h"
 #include "navigation/gpx.h"
 #include "navigation/nav_ids.h"
